move graph parameters to graph module
